[MCCAS][nfc] Add helper classes for writing DIE structures
This commit introduces functions and classes that understand the binary
format used when writing CAS blocks for DIEs.

Note that all the helpers placed in `MCCASDebugV1.h` are meant to be
functions that do not depend on the CAS abstractions. That file
understands the binary formats used but has no knowledge of the CAS.

We add a base class for writing to streams backed by a vector
(`DataWriter`). It is specialized for the three kinds of blocks we are
going to create: Abbreviation entries, Distinct data, and DIE data.