[clang][modules][deps] Including module maps are affecting
With this patch, we mark module maps that include an affecting `extern` module map as also affecting. This is a generalization of D137197: now we don't require the importing module map to describe parent of the extern module.

Depends on D137198.

Reviewed By: Bigcheese

Differential Revision: https://reviews.llvm.org/D137206

This is a combination of upstream commits 50fd6604b76e5fcf5d481a4a551efd50398d9abe and d3804b88e96e40300ec9993ed691bf2184932e29.