[Index] Use correct `printName` when generating a `MSGuidDecl` USR
This is stable/20220421 specific - do not cherry-pick to next.

On stable/20220421, `printName(raw_ostream)` is virtual. But next has
changed that method to be non-virtual and to instead call `printName`
with a `PrintingPolicy`.

This is a fairly wide-reaching change and would require pulling in a
bunch of changes. Just use the virtual `printName` directly instead.