[clang][cas] Map toolchain path based on resource-dir insted of argv0
Use the resource-dir to determine the path of the 'toolchain' for
mapping purposes. In the rare case that this does not match the
compiler, the resource directory is where the input files are more
likely to come from. It also lets us simplify the API since the resource
directory is already in the CompilerInvocation.