Add an SBExpressionOption to allow enabling the "pc macro" transform for playgrounds
The Swift compiler has a transform called "pc macro" that inserts calls to `playground_pc_before()` and `playground_pc_after()` before and after each range of code the PC visits (typically each line of source code).  This transform is performed by calling `swift::performPCMacro()` on the parsed source file expression.

This is analogous to the "playground transform" (and can indeed be combined with `swift::performPlaygroundTransform()`, with the restriction that `swift::performPCMacro()` has to be called _before_ `swift::performPlaygroundTransform()` is called).

This change allows this transform to be controlled in the option struct, which already has things like `GetPlaygroundTransformEnabled()` and `GetREPLEnabled()`, and to call `swift::performPCMacro()` when it's set.  To maintain existing behaviour, this option should default to not being enabled.

rdar://103090524