[mccas debug] Split debug info on class/struct types
This commit makes it so that we create a CAS block per structure/class
entry in the debug information section.

To achieve this, we look for the corresponding tags (TAG_class_type and
TAG_structure_type), but only split those DIEs that do not have a
AT_declaration attribute. This is done to prevent the creation of small
CAS blocks that don't contain the complete type definition.