[clang/cas] Introduce mechanism for caching diagnostics with full source location fidelity
It serializes the subset of `SourceManager` state for the diagnostic
source locations (`FileID`s, macro expansions, etc.) and replays diagnostics by
materializing their `SourceLocation`s; that way the diagnostic consumers have
full access to the include stack, expanded macros, etc.

This allows us to only cache the diagnostic information that is essential for
the compilation and ignore the various ways that diagnostics can be rendered.
And since diagnostic rendering runs in the context of the "local" compilation
we can also remap the canonical file paths of diagnostics to their original sources.

rdar://103448355