[mccas] Use appropriate data stream when extracting Form size
To decide the length of a Form using variable-length encoding, we must
peek at the encoded data. Prior to this commit, we would always use the
"DIEData" array for that purpose, but it is incorrect to do so as the
form data may have been placed in the "DistinctData" array.