[mccas] Use LEB encoding for DWARF_FORM_ref4
These forms account for about 20% of the total size of the CAS. Since
they are stored in the distinct data block, deduplication does not
improve their total contribution to the storage.

As such, this patch changes the encoding of `ref4`s prior to storing
them to the CAS, restoring them to their prior encoding upon
materialization.  This is accomplished by introducing a new form
(`ref4_cas`) to use instead of `ref4`.

The value of the new form was chosen so that it still fits in a single
byte when LEB encoded, and so that it was as far as possible from the
standard forms. We skipped 0x7f because some tests use that value to
test the compiler's behavior when it encounters an unknown form.

Special care needs to be taken so that this change does not affect the
abbreviation section, otherwise preserving the object file byte by byte
may be challenging. "Special care" here means replacing _all_ `ref4`s
with a different form, not just a subset of them.