Do not zero out addends in target asm writer.
This is the first step towards rewriting the logic for how relocation addends are to be handled in MCCAS. Previously, we would zero out the addends in the target asm backend, then store those addends away, and either patch them up in the target object writer or write them to the CAS. This was inefficient because we were storing the addends twice in the CAS, once as the actual addends, and another time as zero-ed out values in the cas blocks themselves. We also had to change a lot of target specific code to achieve this, which would not go down well upstream.