Handle relocation addends in the debug_line section.
To handle relocation addends in the line table section, we can use the line table opcodes to tell us where the addends are in the line table, the DW_LNE_set_address and DW_LNS_fixed_advance_pc opcodes are ones that take a relocation addend as an operand. We can then copy those addend values and store them in the DistinctDebugLineRef. The rest goes into a regular DebugLineRef.