Remove -fcas-friendly-debug-info option and LLVM IR Metadata for cas friendliness, and replace it with an -mllvm option in DwarfDebug.cpp
In an effort to make merging into the stable apple/llvm-project branch easier, and also making cas friendly debug info generation not depend on the driver, this patch removes the old -fcas-friendly-debug-info= option and replaced it with an mllvm option in DwarfDebug.cpp.

This is better because the cas friendly debug info is only generated in the backend and doesn't have anything to do with the compile unit, so the driver should not have any knowledge of this.