Fix the run locker setting for async launches that don't stop at the initial stop.  The code was using PrivateResume when it should have used Resume.
This was allowing expression evaluation while the target was running,
and though that was caught a litle later on, we should never have gotten
that far.  To make sure that this is caught immediately I made an error
SBValue when this happens, and test that we get this error.

Differential Revision: https://reviews.llvm.org/D144665

(cherry picked from commit a92f7832f35c6c4792d8693e724c19802da75b36)