Copy object dump code for mccas into llvm-cas-dump
llvm-cas-object-format is the tool that is used to dump cas statistics, we do not want to merge llvm-cas-object-format into the next branch, but we need the ability to dump cas statistics in the next branch. This patch copies some of the code needed to do cas stastitics dumping into llvm-cas-dump.