[LLDB] Regnerate the static python bindings
This is so that the new APIs added in 55a363fea18b20a81e8ebb9518997a3bda602f32
are available using the static bindings.

Unfortuantely the build of LLDB was broken at the time of committing
this but it looks like the SWIG bindings were successfully regenterated.

The size of the diff is a little surprising though...

rdar://105606978