Fix bug with materialization when one mc:debug_string block is created for the whole string section.
The option to split the debug string section into multiple cas blocks
was already present. However, it had a bug when materializing the object
file from the CAS, While materializing an object file, the code always
adds a null byte after the contents of the mc:debug_string cas object,
this is fine when one cas object is created per string, because the cas
creation splits on the null byte and doesn't store it, but it adds an
extra null byte when one cas object is created for the entire string
section. This bug fix addresses that issue.