[clang][cas] Fix dependency directive lookup for prefix-mapped paths
Currently we are failing to find dependency directives for prefix-mapped
paths in the scanner, such as -include options.

Match how ivfsoverlay marks remapped paths. We want to use the external
(remapped) path when we lookup directives; currently the way to do that
is getName() combined with ExposesExternalVFSPath; if in the future we
add explicit API for this we can drop the status flag.