Fix cas-friendly line table emission in llvm
When using the option -mllvm -cas-friendly-debug-info, we expect the
line table to be emitted in such a way, that it is split-able in the CAS
by every function's line table contribution. This was achieved by
emitting a DW_LNE_end_sequence at the end of the function's contribution
to the line table. This however, had some issues.

Namely, there was an extra DW_LNE_end_sequence being emitted at the end
of the line table.

There was also the case where because of DW_LNE_end_sequence denoting
a reset of the state machine. If the file number of the next line table
is 1, then a new DW_LNS_set_file was not emitted, to fix this and
maximize deduplication, a DW_LNS_set_file is always emitted after a
DW_LNE_end_sequence.

These issues have been addressed along with changes to associated tests.