[clang][cas] Fix cache poison error with diagnostic options
Avoid emitting DIAGNOSTIC_OPTIONS block when caching. In cached builds,
there is no benefit to emitting diagnostic options into the PCM/PCH,
since we know it can never mismatch. Remove this block, because it can
contain volatile options such as fmessage-length that should not have
any impact on the output when using cached diagnostics.

Also add a test for -ferror-limit: while it is okay to clear this option
it's important to test it since it relies on the fact we are not caching
compilation failures. Unlike the other "formatting" options, this
affects the set of diagnostics passed to the diagnostic consumer.

rdar://109285136