[Clang] Fix crash when emitting diagnostic for out of order designated initializers in C++
In C++ we are not allowed to use designated initializers to initialize fields
out of order. In some cases when diagnosing this we are crashing because we are
not indexing correctly and therefore going out of bounds.

This fixes: https://github.com/llvm/llvm-project/issues/63605

Differential Revision: https://reviews.llvm.org/D154675

(cherry picked from commit c9ef33e1d8a8aeb68a18f24af6d9fc9ab4ecf257)

Conflicts:
	clang/docs/ReleaseNotes.rst
	clang/test/SemaCXX/cxx2a-initializer-aggregates.cpp