Replace hand-rolled SharedMutex with C++17 std::shared_mutex (NFCI)
We are getting reports where LLDB is hanging in try_lock().  This
patch makes the code easier to reason about by using standard
primitives, and avoids the pitfall of the copy constructor that the
old code was still allowing. While this may not fix the hangs, it
should certainly make it easier to diagnose.

rdar://112025620