Fix a race condition.
The reader lock must be acquired before retrieving a scratch context,
otherwise another thread could still tear it down between the time it
was created and the creation of the reader object.

There are other shortcomings in GetSwiftScratchContext() that are
documented, but not addressed, in this patch to make it easier to review.