Don't run ObjectFile/PECOFF unit tests on Darwin
A change was cherrypicked adding this test
https://github.com/apple/llvm-project/pull/7186
which uses a keyword that yaml2pecoff on this
branch did not understand.  There was an attempt
to backport the necessary changes via
https://github.com/apple/llvm-project/pull/7189
but it's still failing on the PR bots.  Skip the
test on Darwin for this branch only.

(cherry picked from commit 89d973603acb810aa3bf9d7dd7197784258b89c4)