Prepopulate the CompilerInvocation with the host plugin search path.
Generally, LLDB takes compiler plugin search paths from serialized
options in Swift modules, however, for the REPL and Playgrounds it is
useful to always add the host toolchain plugins by default.

Unfortunately this is a property of the toolchain lldb is installed
with, so this is not really testable from within an uninstalled build.

rdar://112122752