[llvm/CAS] Introduce `ObjectStore::isMaterialized()`
This returns true if the object is directly available from the local CAS, for
implementations that have this kind of distinction. For such implementations it's useful
to be able to identify whether objects need to be downloaded from the distributed cache.

(cherry picked from commit c69249fc78f6451082f16f0d0b796c512ac750cf)