[Serialization] Do less redundant work computing affecting module maps (#66933)
We're traversing the same chains of module ancestors and include
locations repeatedly, despite already populating sets that can detect
it!

This is a problem because translateFile() is expensive. I think we can
avoid it entirely, but this seems like an improvement either way.

I removed a callback indirection rather than giving it a more
complicated
signature, and accordingly renamed the lambdas to be more concrete.

(cherry picked from commit 0f05096540bc90125df47983d7013dd440617055)