[dsymutil] Remove paper trail warnings
Remove the paper trail warning from dsymutil and the DWARF linker. The
original purpose of this functionality was to leave a paper trail in the
output artifact about missing object files. The current implementation
however has diverged and is the source of a pretty serious bug:

  - In the debug map parser, missing object files are not the only
    warnings we emit. When paper trail warnings are enabled, all of them
    end up in the dSYM which wasn't the goal.
  - When warnings are associated with a object file in the debug map, it
    is skipped by the DWARF linker. This only makes sense if the object
    file is missing and is obviously incorrect for any other type of
    warning (such as a missing symbol).

The combination of the two means that we can generate broken DWARF when
the feature is enabled. AFAIK it was only used by Apple and nobody I
spoke to has relied on it, so rather than fixing the broken behavior I
propose we remove it.