Don't update the triple when loading additional dylibs in the expression evaluator
This is a surgical fix for the 5.9 branch to address a design
problem. When a new Swift module is imported in the expression
evaluator, LLDB deserialized the compiler invocation from it and the
way the code was written it replaced the existing compiler invocation
instead of just adding additional search paths. This patch fixes this
by adding a paths-only mode that uses a temporary compiler invocation
object.

rdar://115747362