Add SwiftLanguageRuntime support for implicit enum members.
To verify this I also added a new assertion that replaces the // FIXME
unimplemented comments sprinkled over the code. This uncoverd a bunch
mre bugs related to implicit fields that are also fixed in this
patch. One issue with nested enums is described in a comment but not
yet fixed.